Add Jest tests
This is a pretty basic set of tests. It makes sure mocked fetch requests get called and resolved as we'd expect, and it snapshots the request that the generated code sends to Fetch.
This format will let us test both the logic we need in our own wrapper code, *and* it will let us get a better idea about what changes a different generator might introduce.

One or two other fun features:
1. the tsconfig now includes all TS files, not just the ones in the
   top-level of src (and their imports).
2. absolute imports work now.