Don't have TypeScript do what ESLint should do
The generator generates not-very-lint-safe code (it has unused vars, for example), but TypeScript still needs to know about it.
For things that aren't related to types, we should use ESLint since it's able to ignore parts of our codebase. Since TypeScript needs to know about any code that we use at all, it shouldn't act as our linter.