Don't ignore /build when publishing
We used to publish /build directly, but that's not super conventional —  it means having to put package.json, license files, etc. in the build folder.
It's a lot more common to just publish the whole package and specify the root file in package.json, but to do that we'll need a separate npmignore (.gitignore is used by default) to ensure that out build folder actually gets published.