feat: Expose access to the generated API file
We generate requests from endpoints directly, but they're not always as easy to use as we'd like. For the most commonly-used endpoints, we expose resources that are cleaner and easier to use. For everything else, we need to initialize an API object for that particular API group. Currently, there's no convenient way to do that; you have to make your own API config, like we do in the Octane singleton, and initialize it yourself. It would be a lot easier if you could get access to the API config object itself. This is mostly to speed up ease of use internally (we use octane-node to power a few of our services), since the ergonimics of the SDK matter less to us.